build: keep vc_core Qt-free by splitting VCCollection into data + signaling
Move VCCollection's plain data + I/O (collections map, points map, json
load/save, addPoint, removePoint, etc.) into a new PointCollections
class in vc_core (vc/core/PointCollections.hpp + PointCollections.cpp).
VCCollection in vc_ui shrinks to a tiny QObject subclass that overrides
PointCollections's protected on*() hooks to emit Qt signals.

Tracer + flattening + inpaint + surface_metrics sources merge into
vc_core. CLI tools that previously used VCCollection switch to
PointCollections — none of them connect to signals, they only mutate /
serialize. Result: libvc_core.so has zero Qt in its DT_NEEDED.

Drops UDataManipulateUtils (dead code).
